// src/generated/meta/Line.ts
export const Lines = [
  {
    "id": "1",
    "name": "1",
    "modeName": "bus",
    "disruptions": [],
    "created": "2025-06-24T15:15:34.557Z",
    "modified": "2025-06-24T15:15:34.557Z",
    "lineStatuses": [],
    "routeSections": [],
    "serviceTypes": [
      {
        "name": "Regular",
        "uri": "/Line/Route?ids=1&serviceTypes=Regular"
      }
    ],
    "crowding": "Unknown"
  },
// ...
] as const;